Mod.AI/requirements.txt
discord.py
aiosqlite
openai